feat(credential): add cached and resolver based did doc resolution
