+++ packages/vc-verification/src/index.ts
@@ -1,4 +1,6 @@
 export * from './issuer-verification';
-export * from './credential-resolver';
+export * from './Ipfs-credential-resolver';
 export * from './role-resolvers';
 export * from './role-definition-resolver';
+export * from './cached-diddoc-resolution';
+export * from './credential-resolver';
