+++ packages/vc-verification/src/models/credential-resolver-settings.ts
@@ -1,6 +1,13 @@
 /**
- * Storage settings to fetch issuer's credential
+ * Storage settings to fetch issuer's credential from IPFS
  */
-export interface CredentialResolverSettings {
+export interface CredentialResolverIpfsSettings {
   ipfsUrl: string;
+}
+
+/**
+ * Storage settings to fetch issuer's credential using VUI APIs
+ */
+export interface VuiIssuerApiCredentialResolverSettings {
+  issuersRequestUrl: string;
 }
